Refactor: Moved to->from gpu functions to common_diffraction.py
